Off canvas adjustments
Positioned the off canvas, off canvas by translateX(-100%) where adding the sidebar-visible class via .on(click) should remove that and make it visible. Mastfoot has a display of none in prep for this. Original menu will need something like this too. Increased the sidebar font-size as well as the line-height.